package/libpthsem: correctly handle --{en,dis}able-debug arguments on configure
The pthsem package was incorrectly handling --{en,dis}able-debug. If either of
the two were provided, it always assumed that meant that debugging should be
enabled, which resulted in *extremely* noisy debug logging (e.g. this can be
observed by generating a GPG keypair).

This patch makes pthsem handle the argument correctly.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>