gdb: unbreak build when using --with-expat
The expat library *was* being built first, but the configure script was trying
to link against /usr/lib/libexpat.so on the *host* instead of in the staging
directory. This likely slipped through the cracks because people were doing
buildroot builds which had the same architecture as the host (and thus the link
to the system libexpat.so would succeed).

Signed-off-by: Steven Noonan <steven@uplinklabs.net>